Adding items to wishlist

+++ hebberige-bruiden/App_Code/Entities/Item.cs
@@ -2,6 +2,6 @@
 {
     public int Id { get; set; }
     public string ItemName { get; set; }
-    public int ItemPrice { get; set; }
+    public double ItemPrice { get; set; }
     public string ItemDesc { get; set; }
 }
